+++ tests/test_misc.py
@@ -30,3 +30,26 @@ def test_gap_function_re():
         function ( a, b )
     ''')
     assert m and m.group().strip() == 'function ( a, b )'
+
+
+def test_lazy_function_as_argument():
+    """
+    Regression test for bug with lazy functions.
+
+    When a lazy function is used as an argument to another GAP function, ensure
+    that the function is initialized.
+    """
+
+    @gap.gap_function
+    def OnPoints(omega, g):
+        """
+        Just a wrapper for OnPoints to demonstrate the bug.
+
+        function(omega, g)
+            return OnPoints(omega, g);
+        end;
+        """
+
+    G = gap.Group(gap.eval('(1,2,3)'), gap.eval('(2,3,4)'))
+    O = gap.Orbit(G, 1, OnPoints)
+    assert O == [1, 2, 3, 4]