+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+v0.1.0a1 (unreleased)
+---------------------
+
+* Fixed a bug when using functions defined with ``gap.gap_function`` as
+  arguments to another GAP function before they have been called once.
+
+
 v0.1.0a0 (2021-01-26)
 ---------------------